+++ packages/plugin-svelte-query/src/generators/__snapshots__/postAsQuery.ts
@@ -19,6 +19,7 @@ async function updatePetWithForm(petId: UpdatePetWithFormPathParams["petId"], da
  export function updatePetWithFormQueryOptions(petId: UpdatePetWithFormPathParams["petId"], data?: UpdatePetWithFormMutationRequest, params?: UpdatePetWithFormQueryParams, config: Partial<RequestConfig<UpdatePetWithFormMutationRequest>> = {}) {
     const queryKey = updatePetWithFormQueryKey(petId, data, params);
     return queryOptions({
+        enabled: !!(petId),
         queryKey,
         queryFn: async ({ signal }) => {
             config.signal = signal;