+++ packages/plugin-vue-query/src/generators/__snapshots__/findInfiniteByTagsCursor.ts
@@ -22,7 +22,8 @@ async function findPetsByTags(headers: FindPetsByTagsHeaderParams, params?: Find
     const queryKey = findPetsByTagsInfiniteQueryKey(params);
     return infiniteQueryOptions({
         queryKey,
-        queryFn: async ({ pageParam }) => {
+        queryFn: async ({ signal, pageParam }) => {
+            config.signal = signal;
             if (params) {
                 params["pageSize"] = pageParam as unknown as FindPetsByTagsQueryParams["pageSize"];
             }